Likes. Made Like method internal in MeetingComment aggregate. Rewrote tests of liking for MeetingCommentLikeService only.
